requirements.txt
jira==2.0.0
python-dotenv==0.10.3
